Split of purchases and products
This paves the way for purchases on desktop.
