+++ src/platforms/android/androidiaphandler.h
@@ -19,7 +19,7 @@ class AndroidIAPHandler final : public IAPHandler {
 
  protected:
   void nativeRegisterProducts() override;
-  void nativeStartSubscription(Product* product) override;
+  void nativeStartSubscription(ProductsHandler::Product* product) override;
   void nativeRestoreSubscription() override;
 
  private:
