+++ src/platforms/wasm/wasmiaphandler.h
@@ -17,7 +17,7 @@ class WasmIAPHandler final : public IAPHandler {
 
  protected:
   void nativeRegisterProducts() override;
-  void nativeStartSubscription(Product* product) override;
+  void nativeStartSubscription(ProductsHandler::Product* product) override;
   void nativeRestoreSubscription() override;
 };
